chore: minor re-organization of integration tests and utils
This commit just attempts to re-organize the int test file hierarchy
a bit to make it easier to navigate in the future. Of note:

* Make a subdir for 'cache' integration tests so we can easily separate
  topic and other int tests in the future
* Move the SimpleCacheClient int tests into an `old` subdir to make it
  obvious they are going away
* Move the `cache_test_state` stuff up into the existing `test-util`
  dir, where the doctest helpers are defined, to consolidate all the
  utility/helper code into the same spot.
