+++ Cargo.toml
@@ -54,4 +54,3 @@ uuid = { version = "1", features = ["v4"] }
 futures = "0"
 anyhow = "1"
 momento-test-util = { path = "test-util" }
-once_cell = "1.19.0"
