+++ test-util/Cargo.toml
@@ -10,3 +10,4 @@ anyhow = "1.0.68"
 tokio = { version = "1.25.0", features = ["full"] }
 uuid = { version = "1.2.2", features = ["v4"] }
 scopeguard = "1.1.0"
+once_cell = "1.19.0"
