+++ tests/cache/mod.rs
@@ -1,0 +1,1 @@
+mod sorted_set_tests;
