+++ tests/tests.rs
@@ -1,0 +1,8 @@
+// In order to run tests in nested directories, they must be referenced here.
+mod cache {
+    mod sorted_set_tests;
+}
+
+mod old {
+    mod integration_test;
+}
